requirements.txt
appdirs==1.4.4
attrs==23.1.0
audioread==3.0.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.1.0
cvxopt==1.3.1
cycler==0.11.0
decorator==5.1.1
enum34==1.1.10
fonttools==4.42.0
future==0.18.3
idna==3.4
jams==0.3.4
joblib==1.3.1
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
kiwisolver==1.4.4
lazy_loader==0.3
librosa==0.10.0.post2
llvmlite==0.40.1
matplotlib==3.7.2
midi2audio==0.1.1
mido==1.3.0
mir-eval==0.7
msaf==0.1.80
msgpack==1.0.5
numba==0.57.1
numpy==1.24.4
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pooch==1.6.0
PyAudio==0.2.13
pycparser==2.21
pydub==0.25.1
pygame==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytube==15.0.0
pytz==2023.3
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.12.2
simpleaudio==1.0.4
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.3.5
threadpoolctl==3.2.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vmo==0.30.5
google-api-python-client=2.95.0
streamlit=1.25.0
python-dotenv=1.0.0